Dev (#22)
* Create .gitignore

* Update movie.py

* argument to select a target stream, logging improvements

* bugfix

* log ffmpeg output to console only if console handler is debug or there is an warning

* muxing info logs

+++ aebn_dl/movie.py
@@ -18,12 +18,40 @@ from tqdm import tqdm
 
 class Movie:
     def __init__(self, url, target_height=None, start_segment=None, end_segment=None, ffmpeg_dir=None, scene_n=None, output_dir=None, work_dir=None,
-                 scene_padding=None, log_level="INFO", proxy=None, proxy_metadata_only=False, download_covers=False, overwrite_existing_files=False,
+                 scene_padding=None, log_level="INFO", proxy=None, proxy_metadata_only=False, download_covers=False,
+                 overwrite_existing_files=False, target_stream=None,
                  keep_segments_after_download=False, aggressive_segment_cleaning=False,
                  resolution_force=False, include_performer_names=False, segment_validity_check=False, keep_logs=False):
+        """
+        Parameters:
+        - url: The URL of the movie.
+        - target_height: The desired height of the movie.
+        - start_segment: The starting segment of the movie.
+        - end_segment: The ending segment of the movie.
+        - ffmpeg_dir: The directory for FFMPEG.
+        - scene_n: The scene number of the movie.
+        - output_dir: The directory for output.
+        - work_dir: The directory to store temp files in.
+        - scene_padding: Padding in seconds.
+        - log_level: Logging level (default is "INFO").
+        - proxy: Proxy.
+        - proxy_metadata_only: Use proxy for metadata only.
+        - download_covers: Flag to download covers.
+        - overwrite_existing_files: Flag to overwrite existing files.
+        - target_stream: Download only the specified stream.
+        - keep_segments_after_download: Flag to keep segments after download.
+        - aggressive_segment_cleaning: Flag for aggressive segment cleaning.
+        - resolution_force: Flag to force resolution.
+        - include_performer_names: Flag to include performer names in output file name.
+        - segment_validity_check: Flag for segment validity check.
+        - keep_logs: Flag to keep logs.
+        """
+
         self.movie_url = url
         self.output_dir = output_dir or os.getcwd()
         self.work_dir = work_dir or os.getcwd()
+        self.audio_stream = None
+        self.video_stream = None
         self.target_height = target_height
         self.resolution_force = resolution_force
         self.include_performer_names = include_performer_names
@@ -33,28 +61,43 @@ class Movie:
         self.scene_n = scene_n
         self.download_covers = download_covers
         self.overwrite_existing_files = overwrite_existing_files
+        self.target_stream = target_stream
         self.aggressive_segment_cleaning = aggressive_segment_cleaning
         self.keep_segments_after_download = keep_segments_after_download
         self.scene_padding = scene_padding
         self.log_level = log_level
         self.keep_logs = keep_logs
         self.segment_validity_check = segment_validity_check
-        self.stream_map = []
         self.proxy = proxy
         self.proxy_metadata_only = proxy_metadata_only
         self._logger_setup(log_level)
 
+    class _Media_stream:
+        def __init__(self, human_name, id, movie_work_dir):
+            self.human_name = human_name
+            self.type = self.human_name[0].lower()
+            self.id = id
+            self.path = os.path.join(movie_work_dir, f"{self.type}_{self.id}.mp4")
+            self.downloaded_segments = []
+
     def _logger_setup(self, log_level):
         logger_name = self.movie_url.split("/")[5] + "_" + str(self.scene_n) if self.scene_n else self.movie_url.split("/")[5]
         movie_logger = logging.getLogger(logger_name)
-        movie_logger.setLevel(log_level)
-        formatter = logging.Formatter('%(asctime)s|%(levelname)s|%(message)s', datefmt='%H:%M:%S')
-        main_handler = logging.StreamHandler()
-        main_handler.setFormatter(formatter)
-        movie_logger.addHandler(main_handler)
+        movie_logger.setLevel(logging.DEBUG)  # Set the logger level to the lowest (DEBUG)
 
+        formatter = logging.Formatter('%(asctime)s|%(levelname)s|%(message)s', datefmt='%H:%M:%S')
+
+        # Console handler with user set level
+        console_handler = logging.StreamHandler()
+        console_handler.setLevel(log_level)
+        console_handler.set_name("console_handler")
+        console_handler.setFormatter(formatter)
+        movie_logger.addHandler(console_handler)
+
+        # File handler with DEBUG level
         file_handler = logging.FileHandler(f'{logger_name}.log')
         file_handler.setLevel(logging.DEBUG)
+        file_handler.set_name("file_handler")
         file_handler.setFormatter(formatter)
         movie_logger.addHandler(file_handler)
 
@@ -67,6 +110,12 @@ class Movie:
 
         self.logger = movie_logger
         self.is_silent = True if movie_logger.getEffectiveLevel() > logging.INFO else False
+
+    def _get_handler_level(self, handler_name):
+        for handler in self.logger.handlers:
+            if handler.name == handler_name: 
+                return handler.level
+        return None
 
     def _delete_log(self):
         for handler in self.logger.handlers:
@@ -88,6 +137,7 @@ class Movie:
         self.logger.info(f"Proxy: {self.proxy}") if self.proxy else None
         self.logger.info(f"Output dir: {self.output_dir}")
         self.logger.info(f"Work dir: {self.work_dir}")
+        self.logger.info(f"Target stream: {self.target_stream}") if self.target_stream else None
         self.logger.info(f"Segment validity check: {self.segment_validity_check}")
         if self.aggressive_segment_cleaning:
             self.logger.info("Aggressive cleanup enabled, segments will be deleted before stream muxing")
@@ -109,27 +159,31 @@ class Movie:
         self._get_new_manifest_url()
         self._get_manifest_content()
         self._parse_streams_from_manifest()
+        self.file_name = f"{self.target_stream}_{self.file_name}" if self.target_stream else self.file_name
         self.file_name += f" Scene {self.scene_n}" if self.scene_n else ""
         if self.include_performer_names:
             if not self.performers and self.scene_n:
                 self.logger.info(f"No performers listed for scene {self.scene_n}")
-            if self.performers:
-                self.file_name += " " + ", ".join(self.performers)
-        self.file_name += f" {self.target_height}p"
+            self.file_name += " " + ", ".join(self.performers) if self.performers else ""
+        self.file_name += f" {self.target_height}p" if self.video_stream else ''
         self.output_path = os.path.join(self.output_dir, f"{self.file_name}.mp4")
         self.logger.info(self.file_name)
-        if self.scene_n:
-            self._calculate_scenes_boundaries()
+        self._calculate_scenes_boundaries() if self.scene_n else None
         self._download_segments()
-        self._join_segments_into_stream()
-        self.logger.debug("muxing streams")
-        self._ffmpeg_mux_streams(*[stream["path"] for stream in self.stream_map])
+        self._join_stream_segmetns()
+        if all([self.audio_stream, self.video_stream]):
+            self.logger.info("Muxing streams with ffmpeg")
+            self._ffmpeg_mux_streams(self.audio_stream.path, self.video_stream.path)
+        else:
+            os.remove(self.output_path) if os.path.exists(self.output_path) else None
+            os.rename(self.audio_stream.path, self.output_path) if self.audio_stream else None
+            os.rename(self.video_stream.path, self.output_path) if self.video_stream else None
+
         self._work_folder_cleanup()
         self.logger.info(Path(self.output_path).as_uri())
         self.logger.info(f"{self.file_name}.mp4")
         self.logger.info("Success!")
-        if not self.keep_logs:
-            self._delete_log()
+        self._delete_log() if not self.keep_logs else None
 
     def _create_new_session(self, use_proxies=True):
         self.session = requests.Session()
@@ -313,16 +367,6 @@ class Movie:
                 return stream_id
             self.logger.debug("Skipping a bad audio stream")
 
-    def _add_stream(self, stream_name, stream_id):
-        stream_type = stream_name[0].lower()
-        new_stream = {
-            'human_name': stream_name,
-            'type': stream_type,
-            'id': stream_id,
-            'path': os.path.join(self.movie_work_dir, f"{stream_type}_{self.movie_id}.mp4")
-        }
-        self.stream_map.append(new_stream)
-
     def _parse_streams_from_manifest(self):
         # Parse the XML manifest
         root = ET.fromstring(self.manifest_content, None)
@@ -331,18 +375,21 @@ class Movie:
         video_streams = self._sort_streams_by_video_height(video_adaptation_sets)
         streams_res_string = " ".join(str(video_stream[1]) for video_stream in video_streams)
         self.logger.info(f"Available video streams: {streams_res_string}")
-        self._add_stream("audio", self._find_best_good_audio_stream(video_streams))
-        if self.target_height == 0:
-            video_stream_id, self.target_height = video_streams[0]
-        elif self.target_height is None:
-            video_stream_id, self.target_height = video_streams[-1]
-        elif self.target_height:
-            video_stream_id = next((sublist[0] for sublist in video_streams if sublist[1] == self.target_height), None)
-            if not video_stream_id and self.resolution_force:
-                raise RuntimeError(f"Target video resolution height {self.target_height} not found")
-            else:
-                video_stream_id, self.target_height = next((sublist for sublist in reversed(video_streams) if sublist[1] <= self.target_height), None)
-        self._add_stream("video", video_stream_id)
+        if self.target_stream and self.target_stream == 'audio' or not self.target_stream:
+            audio_stream_id = self._find_best_good_audio_stream(video_streams)
+            self.audio_stream = self._Media_stream("audio", audio_stream_id, self.movie_work_dir)
+        if self.target_stream and self.target_stream == 'video' or not self.target_stream:
+            if self.target_height == 0:
+                video_stream_id, self.target_height = video_streams[0]
+            elif self.target_height is None:
+                video_stream_id, self.target_height = video_streams[-1]
+            elif self.target_height:
+                video_stream_id = next((sublist[0] for sublist in video_streams if sublist[1] == self.target_height), None)
+                if not video_stream_id and self.resolution_force:
+                    raise RuntimeError(f"Target video resolution height {self.target_height} not found")
+                else:
+                    video_stream_id, self.target_height = next((sublist for sublist in reversed(video_streams) if sublist[1] <= self.target_height), None)
+            self.video_stream = self._Media_stream("video", video_stream_id, self.movie_work_dir)
 
     def _total_number_of_data_segments_calc(self, root, total_duration_seconds):
         # Get timescale
@@ -358,10 +405,12 @@ class Movie:
 
     def _work_folder_cleanup(self):
         if not self.keep_segments_after_download:
-            for stream in self.stream_map:
-                os.remove(stream['path']) if os.path.exists(stream['path']) else None
-            for segment_path in self.segment_file_list:
-                os.remove(segment_path) if os.path.exists(segment_path) else None
+            for stream in [self.audio_stream, self.video_stream]:
+                if stream is None:
+                    continue
+                os.remove(stream.path) if os.path.exists(stream.path) else None
+                for segment_path in stream.downloaded_segments:
+                    os.remove(segment_path) if os.path.exists(segment_path) else None
             self.logger.info("Deleted temp files")
 
         os.rmdir(self.movie_work_dir) if not os.listdir(self.movie_work_dir) else None
@@ -381,37 +430,41 @@ class Movie:
 
         self.logger.info(f"Downloading segments {self.start_segment} - {self.end_segment}")
 
-        for stream in self.stream_map:
-            self.logger.debug(f"Downloading stream {stream['type']} {stream['id']}")
+        for stream in [self.audio_stream, self.video_stream]:
+            if stream == None:
+                continue
+            self.logger.debug(f"Downloading {stream.human_name} stream ID: {stream.id}")
             # downloading init segment
-            init_segment_bytes = self._download_segment(stream['type'], stream['id'],
-                                                        overwrite=self.overwrite_existing_files)
+            init_segment_bytes = self._download_segment(stream.type, stream.id,
+                                                        overwrite=self.overwrite_existing_files,
+                                                        media_stream=stream)
 
             # using tqdm object so we can manipulate progress
             # and display it as init segment was part of the loop
 
             segments_to_download = range(self.start_segment, self.end_segment + 1)
-            download_bar = tqdm(total=len(segments_to_download) + 1, desc=stream['human_name'] + " download", disable=self.is_silent)
+            download_bar = tqdm(total=len(segments_to_download) + 1, desc=stream.human_name.capitalize() + " download", disable=self.is_silent)
             download_bar.update()  # increment by 1
             for current_segment_number in segments_to_download:
-                data_segment_bytes = self._download_segment(stream['type'], stream['id'],
+                data_segment_bytes = self._download_segment(stream.type, stream.id,
                                                             segment_number=current_segment_number,
-                                                            overwrite=self.overwrite_existing_files)
+                                                            overwrite=self.overwrite_existing_files,
+                                                            media_stream=stream)
 
                 if self.segment_validity_check and data_segment_bytes:
                     # slow
                     if not self._is_valid_media(init_segment_bytes + data_segment_bytes):
-                        self.logger.info(f"{stream['type']} {current_segment_number} segment media error")
-                        data_segment_bytes = self._download_segment(stream['type'], stream['id'],
+                        self.logger.info(f"{stream.human_name.capitalize()} segment {current_segment_number} media error")
+                        data_segment_bytes = self._download_segment(stream.type, stream.id,
                                                                     segment_number=current_segment_number,
                                                                     overwrite=True)
                         if not self._is_valid_media(init_segment_bytes + data_segment_bytes):
-                            raise Exception(f"{stream['type']}_{stream['id']}_{current_segment_number} Segment not valid!")
+                            raise Exception(f"{stream.type}_{stream.id}_{current_segment_number} Segment not valid!")
 
                 download_bar.update()
             download_bar.close()
 
-    def _download_segment(self, stream_type, stream_id, save_to_disk=True, segment_number=None, overwrite=False, max_tries=2):
+    def _download_segment(self, stream_type, stream_id, save_to_disk=True, segment_number=None, overwrite=False, max_tries=2, media_stream=None):
         if isinstance(segment_number, int):
             segment_name = f"{stream_type}_{stream_id}_{segment_number}"
         else:
@@ -421,6 +474,7 @@ class Movie:
         segment_path = os.path.join(self.movie_work_dir, f"{segment_name}.mp4")
         if os.path.exists(segment_path) and not overwrite:
             self.logger.debug(f"{segment_name} found on disk")
+            media_stream.downloaded_segments.append(segment_path) if media_stream else None
             with open(segment_path, 'rb') as segment_file:
                 segment_bytes = segment_file.read()
                 segment_file.close()
@@ -445,6 +499,7 @@ class Movie:
                         os.mkdir(self.work_dir)
                     with open(segment_path, 'wb') as f:
                         f.write(response.content)
+                    media_stream.downloaded_segments.append(segment_path) if media_stream else None
                     self.logger.debug(f"{segment_name} saved to disk")
                 return response.content
             elif response.status_code == 404 and segment_number == self.total_number_of_data_segments:
@@ -461,13 +516,14 @@ class Movie:
     def _ffmpeg_mux_streams(self, stream_path_1, stream_path_2):
         cmd = f'ffmpeg -i "{stream_path_1}" -i "{stream_path_2}" -y -c copy "{self.output_path}"'
 
-        if self.logger.getEffectiveLevel() > logging.DEBUG:
+        if self._get_handler_level("console_handler") > logging.DEBUG:
             cmd += " -loglevel warning"
 
         cwd = self.ffmpeg_dir if self.ffmpeg_dir else None
         out = subprocess.run(cmd, shell=True, cwd=cwd, capture_output=True, text=True)
         self.logger.warning(f"ffmpeg stderr: {out.stderr}") if out.stderr else None
         assert out.returncode == 0
+        self.logger.info(f"ffmpeg muxing success")
 
     def _join_files(self, files, output_path, tqdm_desc):
         # concats segments into a single file
@@ -483,18 +539,10 @@ class Movie:
                     os.remove(segment_file_path)
         join_bar.close()
 
-    def _join_segments_into_stream(self):
-        self.segment_file_list = []  # only used for clean up
-        for stream in self.stream_map:
+    def _join_stream_segmetns(self):
+        for stream in [self.audio_stream, self.video_stream]:
+            if stream is None:
+                continue
             # delete old joined stream if exists
-            os.remove(stream['path']) if os.path.exists(stream['path']) else None
-            segment_files = []
-            os.remove(stream['path']) if os.path.exists(stream['path']) else None
-            init_path = os.path.join(self.movie_work_dir, f"{stream['type']}i_{stream['id']}.mp4")
-            segment_files.append(init_path)
-            self.segment_file_list.append(init_path)
-            for num in range(self.start_segment, self.end_segment + 1):
-                data_path = os.path.join(self.movie_work_dir, f"{stream['type']}_{stream['id']}_{num}.mp4")
-                segment_files.append(data_path)
-                self.segment_file_list.append(data_path)
-            self._join_files(segment_files, stream['path'], tqdm_desc=f"{stream['human_name']} segments")
+            os.remove(stream.path) if os.path.exists(stream.path) else None
+            self._join_files(stream.downloaded_segments, stream.path, tqdm_desc=f"{stream.human_name} segments")
